Impl: File receive request

+++ src/database/object/TransferObject.h
@@ -24,13 +24,13 @@ public:
     QString deviceId;
     size_t fileSize;
     size_t skippedBytes;
-    qulonglong requestId;
-    qulonglong groupId;
+    quint32 requestId;
+    quint32 groupId;
     int accessPort;
     Type type;
     Flag flag;
 
-    explicit TransferObject(ulong requestId = 0, const QString &deviceId = nullptr,
+    explicit TransferObject(quint32 requestId = 0, const QString &deviceId = nullptr,
                             const Type &type = Type::Incoming, QObject *parent = nullptr);
 
     bool isDivisionObject();
